+++ src/OpeRe.jl
@@ -2,11 +2,13 @@ module OpeRe
 
 include("network.jl")
 include("transportation.jl")
+include("assignment.jl")
 include("shortestpath.jl")
 include("maximumflow.jl")
 
 import .Network
-import .Transportation 
+import .Transportation
+import .Assignment 
 import .ShortestPath
 import .MaximumFlow 
 
@@ -15,13 +17,15 @@ import .Transportation: TransportationProblem, TransportationResult, balance, is
 import .ShortestPath: ShortestPathResult
 import .Network:  Connection, ShortestPathProblem, MaximumFlowProblem
 import .MaximumFlow: MaximumFlowResult
-
+import .Assignment: AssignmentProblem, AssignmentResult
 
 export TransportationProblem, TransportationResult, balance, isbalanced
 export Connection, ShortestPathResult, MaximumFlowResult 
 export ShortestPathProblem, MaximumFlowProblem
+export AssignmentProblem, AssignmentResult
 
 solve(t::TransportationProblem) = Transportation.solve(t)
+solve(a::AssignmentProblem) = Assignment.solve(a)
 
 function solve(c::Array{Connection, 1}; problem = ShortestPathProblem) 
     if problem == ShortestPathProblem