+++ test/runtests.jl
@@ -2,5 +2,6 @@ using Test
 using OpeRe 
 
 include("testtransportation.jl")
+include("testassignment.jl")
 include("testshortestpath.jl")
 include("testmaximumflow.jl")